+++ extensions/CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Update packages parameter in `PhylumApi.analyze` to `PackageDescriptorAndLockfilePath`
 
+### Added
+
+- Added `getJobStatusRaw` and `checkPackagesRaw` APIs for detailed analysis results
+
 ## 5.5.0 - 2023-07-18
 
 ### Added